+++ test/fixtures/custom-game-tracker.ts
@@ -10,19 +10,19 @@ export default class CustomGameTracker extends GameTrackerBase {
     static override workerModule = import.meta.url;
 
     wins = [0, 0, 0];
-    cntGames = 0;
+    games = 0;
 
     merge(tracker: Tracker) {
         if (!isCustomGameTracker(tracker)) return;
         this.wins[0] += tracker.wins[0];
         this.wins[1] += tracker.wins[1];
         this.wins[2] += tracker.wins[2];
-        this.cntGames += tracker.cntGames;
+        this.games += tracker.games;
         this.time += tracker.time;
     }
 
     trackGame(game: Game) {
-        this.cntGames += 1;
+        this.games += 1;
         switch (game.Result) {
             case '1-0':
                 this.wins[0] += 1;